+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- Removed Fedora integration (DR-3005)
+
 ## [3.0.0] - 2024-09-16
 
 ### Added